Moved UnreachableNodes to graph2
