+++ src/app/graph2/util/NodeGraphUtils.js
