+++ src/app/modules/fsa/panels/analysis/AnalysisPanel.js
@@ -4,7 +4,7 @@ import Config from 'deprecated/config.js';
 import Notifications from 'deprecated/system/notification/Notifications.js';
 import NFAToDFAConversionWarningMessage from 'modules/fsa/notifications/NFAToDFAConversionWarningMessage.js';
 
-import {getUnreachableNodes} from 'modules/fsa2/graph/UnreachableNodes.js';
+import { getUnreachableNodes } from 'graph2/util/NodeGraphUtils.js';
 
 class AnalysisPanel extends React.Component
 {
@@ -36,7 +36,7 @@ class AnalysisPanel extends React.Component
         const currentModule = this.props.currentModule;
         const graphController = currentModule.getGraphController();
         const machineController = currentModule.getMachineController();
-        const props = {graphController: graphController, machineController: machineController};
+        const props = { graphController: graphController, machineController: machineController };
         const messageTag = Config.MACHINE_CONVERSION_MESSAGE_TAG;
 
         //machineController.convertMachineTo("DFA");
@@ -61,7 +61,7 @@ class AnalysisPanel extends React.Component
     canOptimize()
     {
         return (this.optimizeRedundOption && this.optimizeRedundOption.checked) ||
-    (this.optimizeUnreachOption && this.optimizeUnreachOption.checked);
+            (this.optimizeUnreachOption && this.optimizeUnreachOption.checked);
     }
 
     render()
@@ -71,27 +71,27 @@ class AnalysisPanel extends React.Component
 
         const machineType = machineController.getMachineType();
 
-        return <div className={'panel-container ' + this.props.className} id="analysis" ref={ref=>this.container=ref} style={this.props.style}>
+        return <div className={'panel-container ' + this.props.className} id="analysis" ref={ref => this.container = ref} style={this.props.style}>
             <div className="panel-title"><h1>{I18N.toString('component.analysis.title')}</h1></div>
             <div className="panel-content">
 
                 {/*Optimize*/}
-                <div style={{paddingBottom: '0.5em'}}>
-                    <h3 style={{marginBottom: '0'}}>Optimizations</h3>
-                    <div style={{paddingBottom: '0.5em'}}>
+                <div style={{ paddingBottom: '0.5em' }}>
+                    <h3 style={{ marginBottom: '0' }}>Optimizations</h3>
+                    <div style={{ paddingBottom: '0.5em' }}>
                         <div>
-                            <input ref={ref=>this.optimizeUnreachOption=ref} id="opt-unreach" type="checkbox"/>
+                            <input ref={ref => this.optimizeUnreachOption = ref} id="opt-unreach" type="checkbox" />
                             <label htmlFor="opt-unreach">Unreachables</label>
                         </div>
                         <div>
-                            <input ref={ref=>this.optimizeRedundOption=ref} id="opt-redund" type="checkbox" disabled={true}/>
+                            <input ref={ref => this.optimizeRedundOption = ref} id="opt-redund" type="checkbox" disabled={true} />
                             <label htmlFor="opt-redund">Redundant States</label>
                         </div>
                     </div>
                     <button className="panel-button" onClick={this.onOptimizeMachine} disabled={!this.canOptimize()}>Optimize</button>
                 </div>
 
-                <hr/>
+                <hr />
                 {
                     machineType == 'DFA' ?
                         <button className="panel-button" onClick={this.onConvertToNFA}
@@ -112,7 +112,7 @@ class AnalysisPanel extends React.Component
     }
 }
 Object.defineProperty(AnalysisPanel, 'TITLE', {
-    get: function() { return I18N.toString('component.analysis.title'); }
+    get: function () { return I18N.toString('component.analysis.title'); }
 });
 AnalysisPanel.UNLOCALIZED_NAME = 'component.analysis.title';
