+++ src/app/modules/fsa2/machine/FSABuilder.js
@@ -2,7 +2,7 @@ import AbstractMachineBuilder from 'modules/abstract/AbstractMachineBuilder.js';
 import FSA, { EMPTY_SYMBOL, State } from './FSA.js';
 import FSANode from 'modules/fsa2/graph/element/FSANode.js';
 import { EMPTY_CHAR } from 'modules/fsa2/graph/element/FSAEdge.js';
-import {getUnreachableNodes} from 'modules/fsa2/graph/UnreachableNodes.js';
+import { getUnreachableNodes } from 'graph2/util/NodeGraphUtils.js';
 
 export const ERROR_UNREACHABLE_STATE = 'unreachable_state';
 export const ERROR_DUPLICATE_STATE = 'duplicate_state';
@@ -190,7 +190,7 @@ class FSABuilder extends AbstractMachineBuilder
         }
 
         //Check for unreachable nodes
-        const unreachables =  getUnreachableNodes(graph);
+        const unreachables = getUnreachableNodes(graph);
         if (unreachables && unreachables.length > 0)
         {
             warnings.push({
